move envcache types and funcs to separate file
